create voucher

+++ Flutter.Backend.DAL/Domains/Voucher.cs
@@ -12,13 +12,13 @@ namespace Flutter.Backend.DAL.Domains
 
         public int DisCountPercent { get; set; }
 
-        public int DisCountAmount { get; set; }
+        public decimal DisCountAmount { get; set; }
 
         public string ImageVoucher { get; set; }
 
-        public int FromCondition { get; set; }
+        public decimal FromCondition { get; set; }
 
-        public int ToCondition { get; set; }
+        public decimal ToCondition { get; set; }
        
         public string Description { get; set; }
