+++ Flutter.Backend.Service/Models/Dtos/DtoVoucher.cs
@@ -14,11 +14,11 @@ namespace Flutter.Backend.Service.Models.Dtos
 
         public int DisCountPercent { get; set; }
 
-        public int DisCountAmount { get; set; }
+        public decimal DisCountAmount { get; set; }
 
-        public int FromCondition { get; set; }
+        public decimal FromCondition { get; set; }
 
-        public int ToCondition { get; set; }
+        public decimal ToCondition { get; set; }
 
         public string Description { get; set; }
